requirements.txt
click==8.0.1
debugpy==1.4.1
dnspython==2.1.0
flask==2.0.1
itsdangerous==2.0.1
jinja2==3.0.1
markupsafe==2.0.1
python-etcd==0.4.5
urllib3==1.26.6
werkzeug==2.0.1
